requirements.txt
Flask==0.10.1
Jinja2==2.8
MarkupSafe==0.23
Werkzeug==0.10.4
itsdangerous==0.24
six==1.9.0
wsgiref==0.1.2
uwsgi==2.0.12
python-crontab==2.0.1
